requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
sentence_transformers
langchain_huggingface
chromadb
faiss-cpu
langchain_openai
beautifulsoup4
streamlit
langchain_ollama
langchain_groq
